# Reinitialized existing git directory

+++ app/main.py
@@ -10,12 +10,15 @@ def main():
 
     command = sys.argv[1]
     if command == "init":
-        os.mkdir(".git")
-        os.mkdir(".git/objects")
-        os.mkdir(".git/refs")
-        with open(".git/HEAD", "w") as f:
-            f.write("ref: refs/heads/main\n")
-        print("Initialized git directory")
+        if os.path.exists(".git"):
+            print("Reinitialized existing git directory")
+        else:
+            os.mkdir(".git")
+            os.mkdir(".git/objects")
+            os.mkdir(".git/refs")
+            with open(".git/HEAD", "w") as f:
+                f.write("ref: refs/heads/main\n")
+            print("Initialized git directory")
     else:
         raise RuntimeError(f"Unknown command #{command}")
